Restructre cli to have subcommands
This change moves the orginal `helm diff` command to `helm diff upgrade` and adds a `helm diff version` subcommand.
Using the root command is depreacted but still works to maintain backward compatibility.
This change is a preparation to allow for different modes in the plugin (e.g. diffing between revisions).