fix install-binary.sh version specification
This fixes a bug where the --version flag that can be passed to `helm
plugin install` was not working with the install-binary.sh script that
this repository uses to install the binary files for users.

The issue was that the Helm runs install hooks in the current working
directory [1], so the `git describe` command that this repository was
executing was not being executed in the directory of the plugin, and
would thus always fail. The fix here is to use the `-C` flag to git to
specify the directory of the plugin source, ensuring that it will work
correctly regardless of the working directory from which the script is
executed.

Closes https://github.com/databus23/helm-diff/issues/50

[1] https://github.com/kubernetes/helm/blob/401f8bcc187d7a9edeec12c26c64819f105954d5/cmd/helm/plugin.go#L48-L72