Bumping helm dep to v2.11.0
The LDFLAG build metadata for parsing the helm package inside
glide.yaml has to have the next line be "version". It parses
"version" to validate the helm version used to build helm-diff.

Ended up bumping to helm v2.11.0 since that was released 2 weeks
ago and the previous comment mentioned removing it once PR #4499
was released in helm.