env for shebang and add freebsd to install script
Shebang shouldnt hardcode path to /bin/bash since some *nix flavors
dont ship with bash or install it in a different path.

install-binary.sh which is used by the helm plugin system was updated
for a freebsd binary support. This should be functional when the
first release for helm-diff is released for freebsd.