Add --disable-validation for disabling validation against K8s cluster entirely
`--disable-openapi-validation` seems to not work against charts like prometheus-operator because it results in "missing kind" errors due to that it tries to install CRD and CR on first install. That is different than OpenAPI schema validation, which seems rather for detecting invalid fields.

The addition of `--disable-validation` disables the manifests validation against the targeted K8s cluster(again, since Helm 3 `helm template --validate` runs validation against the live cluster so it is relevant), so that we can make use of helm-diff even for cases `--disable-openapi-validation` doesn not work.

Resolves #183