Discover renames based on content
If we rename objects, they will appear as
add and remove actions, which makes it difficult
to assess the semantic delta, which may only be
a simple renaming.

The current implementation creates a diff
between every added & removeed item of the same kind,
which means at worst O(m*n) runtime.
So it should only be enabled if such changes are rare